+++ src/cmd/mod.rs
@@ -5,10 +5,10 @@ pub mod gdb;
 pub mod heap;
 pub mod log;
 pub mod new;
+pub mod print;
 pub mod reset;
-pub mod support;
 
 pub use self::{
     flash::run as flash, gdb::run as gdb, heap::run as heap, log::run as log, new::run as new,
-    reset::run as reset, support::run as support,
+    print::run as print, reset::run as reset,
 };